win32 lfs fix

+++ branches/avidemux_2.1_branch/avidemux/ADM_library/default.h
@@ -59,7 +59,7 @@
 // For win32, useless else
 
 uint64_t ftello_adm(FILE *f);
-uint64_t fseeko_adm(FILE *f,fpos_t off,int whence);
+int fseeko_adm(FILE *f,fpos_t off,int whence);
 
 
 #endif
